refact: CavasCore에서 Edges, Nodes 로직 hook으로 분리
